Allow HTTPS connections
This commit will (hopefully) allow someone to connect to the Hub via
HTTPS. To do this, we can pass a `-https` option in the command-line
when starting a hub. This is a boolean, so will be set to `false` by
default. An example would look like:

```
java -jar selenium-server-standalone.jar -role hub -host localhost -https -port 4444
```

Which will then set the registration URL as `https://localhost:4444/grid/register`